OneClickML/requirements.txt
Flask==1.1.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
mtcnn==0.1.0
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.2
requests==2.25.1
requests-oauthlib==1.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.4.1
tensorflow-estimator==2.4.0
gunicorn>=19.9.0